pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-oracle-agent"
version = "0.1.0"
description = "Autonomous AI Agent for Polymarket"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "web3>=6.15.0",
    "anthropic>=0.19.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1"
]
